some refactoring and some function header comments 	modified:   array_api.py 	modified:   lib.py
